requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.2
faster-whisper>=1.1.1
av>=14.0.0
requests==2.32.3
yt-dlp
google-genai
gunicorn==21.2.0
